requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
streamlit>=1.28.0
plotly>=5.17.0
python-dotenv>=1.0.0
geopy>=2.3.0
joblib>=1.3.0
